storcon: pull client type into shared crate
