+++ storage_controller/client/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod control_api;
